clap detector

+++ clapbot/test.py
@@ -14,7 +14,7 @@ while True:
         BrickPiUpdateValues()
         print BrickPi.Sensor[PORT_1]
         print good(BrickPi.Sensor[PORT_1])
-        print ( 1 - ( int(BrickPi.Sensor[PORT_1]) / 1000 ) )
+        print ( 1 - float( int(BrickPi.Sensor[PORT_1]) / 1000 ) )
 
     except KeyboardInterrupt:
         break
